Fix comic nav not rendering correctly
Lets say a comic had 3 pages and the user clicked the middle one
from the index. It should display the prev and next arrow buttons
but instead it was displaying any.

The prev and next buttons work by figuring out the next or prev
index. Therefore, the logic should check to see if < 1 means the
current index is page 1 so don't show the prev arrow. If next is
greater than the number of pages, don't show the next arrow.